Finish the backend

+++ backend/controllers/goalController.js
@@ -1,9 +1,10 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel')
 
 // Function responsible for Getting Data 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({ user: req.user.id })
   res.status(200).json(goals);
 })
 
@@ -15,6 +16,7 @@ const setGoals = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id
   })
   res.status(201).json(goal)
 })
@@ -26,6 +28,18 @@ const updateGoals = asyncHandler(async (req, res) => {
   if(!goal) {
     req.status(400)
     throw new Error('Goal not found');
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if(!user){
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User not authorized');
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
@@ -41,6 +55,18 @@ const deleteGoals = asyncHandler(async (req, res) => {
     throw new Error('Goal not found');
   }
 
+  const user = await User.findById(req.user.id);
+
+  if(!user){
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User not authorized');
+  }
+  
   await Goal.findByIdAndDelete(req.params.id)
   res.status(200).json({ id : req.params.id})
 })
